Altered TimeStamp to use long value internally to eliminate need for rounding the former double value thus boosting performance.

+++ Platform/TickZoomAPI1.0/EngineInterfaces/Interval.cs
@@ -37,7 +37,7 @@ namespace TickZoom.Api
 			get;
 		}
 		
-		double Period {
+		int Period {
 			get;
 		}
